+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to BlastDesign-AI will be documented in this file.
 
 ## [Unreleased]
 
-## [0.4.0] - 2026-09-23
+## [0.4.0] - 2026-09-24
 
 ### Added